# openapi.yaml
openapi: 3.0.0
info:
  title: Loom Coaching App API
  description: |
    Comprehensive API for the Loom coaching platform, providing secure endpoints for authentication, session management, file handling, and real-time communication between coaches and clients.
    
    ## Authentication
    Most endpoints require authentication via Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    All endpoints are rate-limited for security and performance:
    - Authentication endpoints: 10 requests/minute
    - File uploads: 20 requests/hour
    - General API calls: 100 requests/minute
    
    ## Error Handling
    The API uses consistent error response format:
    ```json
    {
      "success": false,
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: Loom API Support
    email: support@loom-app.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://loom-app.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /sessions/{id}:
    get:
      tags:
        - Sessions
      summary: Get session details
      description: Retrieves detailed information about a specific session
      operationId: getSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      responses:
        '200':
          description: Session details retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Sessions
      summary: Update session
      description: Updates session details. Requires appropriate permissions.
      operationId: updateSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Sessions
      summary: Delete session
      description: Deletes a session. Only admins and session owners can delete.
      operationId: deleteSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Session ID
      responses:
        '204':
          description: Session deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoints
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "12345678-1234-1234-1234-123456789012"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        role:
          type: string
          enum: [client, coach, admin]
          description: User role
          example: "client"
        firstName:
          type: string
          description: User's first name
          example: "John"
        lastName:
          type: string
          description: User's last name
          example: "Doe"
        avatarUrl:
          type: string
          format: uri
          nullable: true
          description: URL to user's avatar image
          example: "https://example.com/avatar.jpg"
        language:
          type: string
          enum: [en, he]
          description: User's preferred language
          example: "en"
        status:
          type: string
          enum: [active, inactive, pending, suspended]
          description: User account status
          example: "active"
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
          description: Last activity timestamp
          example: "2024-01-15T12:00:00Z"
        mfaEnabled:
          type: boolean
          description: Whether MFA is enabled
          example: false
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T12:00:00Z"
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique session identifier
          example: "12345678-1234-1234-1234-123456789012"
        coachId:
          type: string
          format: uuid
          description: Coach user ID
          example: "87654321-4321-4321-4321-210987654321"
        clientId:
          type: string
          format: uuid
          description: Client user ID
          example: "11111111-2222-3333-4444-555555555555"
        title:
          type: string
          description: Session title
          example: "Career Planning Session"
        description:
          type: string
          nullable: true
          description: Session description
          example: "Discussion about career goals and development plan"
        scheduledAt:
          type: string
          format: date-time
          description: Scheduled session time
          example: "2024-01-15T14:30:00Z"
        duration:
          type: integer
          description: Session duration in minutes
          example: 60
        durationMinutes:
          type: integer
          description: Alternative duration field
          example: 60
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, no_show]
          description: Session status
          example: "scheduled"
        meetingUrl:
          type: string
          format: uri
          nullable: true
          description: Virtual meeting URL
          example: "https://meet.example.com/session123"
        notes:
          type: string
          nullable: true
          description: Session notes
          example: "Client showed great progress on goals"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          description: Session rating (1-5)
          example: 5
        coach:
          allOf:
            - $ref: '#/components/schemas/User'
            - type: object
              description: Coach details
        client:
          allOf:
            - $ref: '#/components/schemas/User'
            - type: object
              description: Client details
        createdAt:
          type: string
          format: date-time
          description: Session creation timestamp
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T12:00:00Z"
...
    UpdateSessionRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Session title
          example: "Updated Career Planning Session"
        description:
          type: string
          maxLength: 500
          description: Session description
          example: "Updated discussion about career goals"
        scheduledAt:
          type: string
          format: date-time
          description: Scheduled session time
          example: "2024-01-15T15:30:00Z"
        duration:
          type: integer
          minimum: 15
          maximum: 240
          description: Session duration in minutes
          example: 90
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
          description: Session status
          example: "scheduled"
        meetingUrl:
          type: string
          format: uri
          description: Virtual meeting URL
          example: "https://meet.example.com/session456"
        notes:
          type: string
          maxLength: 2000
          description: Session notes
          example: "Client showed improvement in goal setting"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Validation failed"
        code:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        details:
          type: object
          additionalProperties: true
          description: Additional error details
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error
              value:
                success: false
                error: "Validation failed"
                code: "VALIDATION_ERROR"
                details:
                  issues:
                    - field: "email"
                      message: "Invalid email format"
    Unauthorized:
      description: Unauthorized - Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required. Please provide a valid Bearer token."
            code: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Access denied. Required role: admin"
            code: "FORBIDDEN"
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
...